*** empty log message ***

+++ ebwt.h
@@ -265,14 +265,14 @@ public:
 /**
  * Calculate size of file with given name.
  */
-static inline int fileSize(const char* name) {
+static inline int64_t fileSize(const char* name) {
 	std::ifstream f;
 	f.open(name, std::ios_base::binary | std::ios_base::in);
 	if (!f.good() || f.eof() || !f.is_open()) { return 0; }
 	f.seekg(0, std::ios_base::beg);
 	std::ifstream::pos_type begin_pos = f.tellg();
 	f.seekg(0, std::ios_base::end);
-	return static_cast<int>(f.tellg() - begin_pos);
+	return static_cast<int64_t>(f.tellg() - begin_pos);
 }
 
 // Forward declarations for Ebwt class
